software/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
#psycopg2==2.9.9
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.29


certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
decorator==4.4.2
googletrans==4.0.0rc1
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
langdetect==1.0.9
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
PyAudio==0.2.14
pydub==0.25.1
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1


anyio==4.4.0
cffi==1.16.0
cmake==3.25.0
distlib==0.3.8
filelock==3.14.0
platformdirs==4.2.1
pycparser==2.22
pygame==2.5.2
scipy==1.13.1
sounddevice==0.4.7
srt==3.5.3
threaded==4.2.0
utils==1.0.2
virtualenv==20.26.1
vosk==0.3.45
websockets==12.0



comtypes==1.4.4
itsdangerous==2.2.0
pypiwin32==223
pyttsx3==2.90
pywin32==306
Werkzeug==3.0.3


